requirements.txt
playwright>=1.49.0
google-genai>=1.0.0
python-telegram-bot>=22.0
duckduckgo-search>=7.0.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
lxml>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
